requirements_core.txt
# Core dependencies for face recognition application (without dlib/face-recognition for now)
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
numpy>=1.24.0
Pillow>=9.0.0
pymysql>=1.0.0
mysql-connector-python>=8.0.0
tkVideo>=0.1.0
streamlit>=1.28.0
Django>=4.2.0
requests>=2.28.0
beautifulsoup4>=4.11.0

